move client thread starting after event creation

+++ src/server.cpp
@@ -40,9 +40,11 @@ DWORD WINAPI InteractWithClientThread(LPVOID _args)
 {
     const std::string pipeName = R"(\\.\pipe\os-lab5-pipe)";
     ThreadArgs args = *(ThreadArgs *)_args;
-    size_t id = args.id;
+    size_t threadId = args.id;
     size_t numberOfClients = args.numberOfClients;
     size_t numberOfRecords = args.numberOfRecords;
+    Controller *ctrl = args.ctrl;
+    CRITICAL_SECTION *iocs = args.iocs;
 
     // creating named pipe
     HANDLE pipe = CreateNamedPipeA(
@@ -87,23 +89,46 @@ DWORD WINAPI InteractWithClientThread(LPVOID _args)
         else if (request == 1)
         {
             // reading record
-
             DWORD bytes;
-            int key;
-            ReadFile(pipe, reinterpret_cast<char*>(&key), sizeof(int), &bytes, NULL);
-            std::cout << "Thread received key to read: " << key << "\n";
-            char message[5];
-            message[0] = 1;
-            Sleep(10000);
-            memcpy(&message[1], reinterpret_cast<char *>(&key), 4);
-            WriteFile(pipe, message, 5, &bytes, NULL);
 
+            int key; // employee id
+            ReadFile(pipe, reinterpret_cast<char *>(&key), sizeof(int), &bytes, NULL);
+
+            EnterCriticalSection(iocs);
+            std::cout << "Thread " << threadId << " received key to read: " << key << "\n";
+            LeaveCriticalSection(iocs);
+
+            size_t recordId;
+            bool keyExists = ctrl->idToRecordId(key, recordId); // record id
+
+            if (keyExists)
+            {
+                std::string modifyEventNameTemplate = "Write event ";
+                std::ostringstream modifyEventName(modifyEventNameTemplate);
+                modifyEventName << recordId;
+                HANDLE notBeingModifiedEvent = OpenEventA(EVENT_ALL_ACCESS, FALSE, modifyEventName.str().c_str());
+
+                EnterCriticalSection(iocs);
+                std::cout << "Thread " << threadId << "is waiting for record with key " << key << " to become available\n";
+                LeaveCriticalSection(iocs);
+
+                WaitForSingleObject(notBeingModifiedEvent, INFINITE); // waiting for record to become available
+
+                char message[5];
+                message[0] = 1;
+                Sleep(10000);
+                memcpy(&message[1], reinterpret_cast<char *>(&key), 4);
+                WriteFile(pipe, message, 5, &bytes, NULL);
+            }
+            else
+            {
+            }
         }
         else if (request == 2)
         {
             // modifying record
         }
-        else 
+        else
         {
             std::cerr << "Protocol violation\n";
         }
@@ -160,20 +185,13 @@ int main()
         "Enter number of clients: ",
         "Value must be positive integer\n");
 
-    HANDLE *threads = new HANDLE[numberOfClients];
-    for (size_t i = 0; i < numberOfClients; ++i)
-    {
-        DWORD thread_id;
-        ThreadArgs args;
-        args.id = i;
-        args.numberOfClients = numberOfClients;
-        args.numberOfRecords = numberOfRecords;
-        threads[i] = CreateThread(NULL, 0, InteractWithClientThread, (LPVOID *)(&args), NULL, &thread_id);
-    }
+    // creating critical sections for safe output via stdout
+    CRITICAL_SECTION iocs;
+    InitializeCriticalSection(&iocs);
 
     // creating set of numberOfClients events for every numberOfRecords
     // event will be set if record i is not read by client j
-    HANDLE **notReadEvents = new HANDLE*[numberOfRecords];
+    HANDLE **notReadEvents = new HANDLE *[numberOfRecords];
     for (size_t i = 0; i < numberOfRecords; ++i)
     {
         notReadEvents[i] = new HANDLE[numberOfClients];
@@ -197,6 +215,19 @@ int main()
         notWriteEvents[i] = CreateEventA(NULL, FALSE, TRUE, eventName.str().c_str());
     }
 
+    // starting client interaction
+    HANDLE *threads = new HANDLE[numberOfClients];
+    for (size_t i = 0; i < numberOfClients; ++i)
+    {
+        DWORD thread_id;
+        ThreadArgs args;
+        args.id = i;
+        args.numberOfClients = numberOfClients;
+        args.numberOfRecords = numberOfRecords;
+        args.ctrl = &ctrl;
+        threads[i] = CreateThread(NULL, 0, InteractWithClientThread, (LPVOID *)(&args), NULL, &thread_id);
+    }
+
     // starting clients
     for (size_t i = 0; i < numberOfClients; ++i)
     {
@@ -207,6 +238,7 @@ int main()
     WaitForMultipleObjects(numberOfClients, threads, TRUE, INFINITE);
 
     // freeing memory
+    DeleteCriticalSection(&iocs);
     for (size_t i = 0; i < numberOfRecords; ++i)
     {
         for (size_t j = 0; j < numberOfClients; ++i)
